% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{polmineR-shiny-package}
\alias{polmineR-shiny-package}
\title{Package 'polmineR-shiny'}
\description{
A plugin for the polmineR-package
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\references{
http://polmine.sowi.uni-due.de
}
\keyword{package}

