% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{shinyReferencePartition}
\alias{shinyReferencePartition}
\title{default reference partition for shiny browser apps}
\format{a character vector}
\description{
default reference partition for shiny browser apps
}
\keyword{datasets}

