% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{shinySession-class}
\alias{begin}
\alias{begin,shinySession-method}
\alias{context,shinySession-method}
\alias{dispersion,shinySession-method}
\alias{keyness,shinySession-method}
\alias{kwic,shinySession-method}
\alias{partition,shinySession-method}
\alias{shinySession}
\alias{shinySession-class}
\alias{shutdown}
\alias{shutdown,shinySession-method}
\title{class to manage a shiny session}
\usage{
shinySession()

\S4method{shutdown}{shinySession}(object)

\S4method{partition}{shinySession}(object, ...)

\S4method{kwic}{shinySession}(object, ...)

\S4method{context}{shinySession}(object, ...)

\S4method{keyness}{shinySession}(x, ...)

\S4method{dispersion}{shinySession}(object, ...)
}
\arguments{
\item{object}{the shinySession-object}

\item{...}{further parameters to be passed}

\item{x}{alternative way for the shinySession-object}
}
\description{
class to manage a shiny session

Generate shinySession
}
\section{Slots}{

\describe{
\item{\code{sockets}}{the sockets are stored here}

\item{\code{partitionDir}}{taken from drillingControls}
}}

